requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
pypdf
python-dotenv
langchain-groq
langchain-pinecone
langchain-community
pypdf
langchain-huggingface
langchain-google-genai
-e .